requirements.txt
python-dotenv == 1.0.1
streamlit == 1.31.1
pyrebase == 3.0.27
langchain == 0.3.9
langchain-community == 0.3.8
langchain-experimental == 0.3.3
langchain-groq == 0.2.1